Refactor: 메인페이지 템플릿 파일 통합
