+++ src/routes/URLRouter.js
@@ -1,5 +1,3 @@
-import { DATA } from '../utils/constants'
-
 export function navigate(path) {
   const { pathname } = location
